# Оптимизация времени работы stories#index