--- a/Gemfile
+++ b/Gemfile
@@ -103,6 +103,9 @@ gem "validate_url", "~> 1.0"
 gem "webpacker", "~> 3.5"
 gem "webpush", "~> 0.3"
 
+gem 'newrelic_rpm'
+gem 'prometheus_exporter'
+
 group :development do
   gem "better_errors", "~> 2.5"
   gem "binding_of_caller", "~> 0.8"