--- a/prometheus/docker-compose.yml
+++ b/prometheus/docker-compose.yml
@@ -16,10 +16,11 @@ services:
     ports:
         - '9090:9090'
   grafana:
-    image: grafana/grafana:3.0.0-beta7
+    image: grafana/grafana:5.1.0
     environment:
       - GF_SECURITY_ADMIN_PASSWORD=pass
     depends_on:
       - prometheus
     ports:
      - "3030:3000"
+    user: "104"