--- a/prometheus/prometheus.yml
+++ b/prometheus/prometheus.yml
@@ -1,0 +1,11 @@
+global:
+  scrape_interval: 5s
+  external_labels:
+    monitor: 'my-monitor'
+scrape_configs:
+  - job_name: 'prometheus'
+    target_groups:
+      - targets: ['localhost:9090']
+  - job_name: 'devdev'
+    target_groups:
+      - targets: ['dockerhost:9394']