## Description
Инструменты мониторинга [показывают](https://www.dropbox.com/s/yg0n1iil9rq8q3o/%D0%A1%D0%BA%D1%80%D0%B8%D0%BD%D1%88%D0%BE%D1%82%202019-03-29%2003.03.51.png?dl=0), что самой горячей точкой является главная страница, `StoriesController#index`.
В частности, заметное время занимает рендеринг partial-ов `_single_story.html.erb`.

Нагрузка имитировалась с использованием инструмента `siege` с произвольным набором url'ов 
<pre>
siege -c 5 -t120s -f load_testing/siege_urls.txt
</pre>
Но получение данных для оценки оптимизаций занимает довольно много времени, что не очень удобно.

## Benchmark
Для оценки целесообразности вносимых изменений используем `benchmark` с помощью `ab`
<pre>
ab -n 10 -c 1 127.0.0.1:3000/
</pre>
Это позволит получить быстрый `feedback-loop`

### Before
Данные до внесения изменений.
<pre>
Time taken for tests:   2.576 seconds
Complete requests:      10
Failed requests:        0
Total transferred:      1798040 bytes
HTML transferred:       1787330 bytes
Requests per second:    3.88 [#/sec] (mean)
Time per request:       257.626 [ms] (mean)
Time per request:       257.626 [ms] (mean, across all concurrent requests)
Transfer rate:          681.57 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.0      0       0
Processing:   220  258  25.3    266     297
Waiting:      220  257  25.3    265     297
Total:        220  258  25.3    266     297

Percentage of the requests served within a certain time (ms)
  50%    266
  66%    268
  75%    272
  80%    276
  90%    297
  95%    297
  98%    297
  99%    297
 100%    297 (longest request)
</pre>

### After
Данные после добавления кэширования в `app/views/stories/_main_stories_feed.html.erb`

<pre>
Time taken for tests:   1.460 seconds
Complete requests:      10
Failed requests:        0
Total transferred:      1779180 bytes
HTML transferred:       1768470 bytes
Requests per second:    6.85 [#/sec] (mean)
Time per request:       146.045 [ms] (mean)
Time per request:       146.045 [ms] (mean, across all concurrent requests)
Transfer rate:          1189.69 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.0      0       0
Processing:   119  146  27.5    140     186
Waiting:      119  146  27.5    139     186
Total:        119  146  27.5    140     186

Percentage of the requests served within a certain time (ms)
  50%    140
  66%    142
  75%    183
  80%    184
  90%    186
  95%    186
  98%    186
  99%    186
 100%    186 (longest request)
</pre>

Видно, что кэширование помогло ускорить запрос.

Но есть одна проблема. В кэшируемой части есть вывод кол-ва лайков и комментариев. С высокой вероятностью это часто изменяемые величины 😞 
Есть два решения:
- оставить кэш валидным на допустимо короткий период (5 минут)
- вынести каунтеры в отдельный эндпоинт и дергать его асинхронно, обновляя значения отдельно от загрузки статей.
В связи с отсутствием данных по реальной частоте изменения кол-ва лайков и комментариев, оставим пятиминутный кэш.

[График](https://www.dropbox.com/s/eciwtig062lsiqv/%D0%A1%D0%BA%D1%80%D0%B8%D0%BD%D1%88%D0%BE%D1%82%202019-03-29%2002.06.46.png?dl=0) мониторинга на котором виден эффект оптимизации