@darialapina 2019-03-29
Вам спасибо за интересный материал  🙂 
На самом деле, пока игралась с инструментами, делала и разные изменения и разные параметры для нагрузки и для ab (кол-во запросов, конаренси), и для siege (конаренси, продолжительность нагрузки).